requirements.txt
torch
torchvision
numpy
matplotlib
tqdm
Pillow
wandb
datasets
huggingface_hub
